[WIP] Align new settings panels with other cards
ref https://linear.app/ghost/issue/PLG-374/align-new-settings-panel-settings-with-other-cards
