+++ packages/koenig-lexical/src/components/ui/ButtonGroup.old.stories.jsx
@@ -2,12 +2,12 @@ import ImgFullIcon from '../../assets/icons/kg-img-full.svg?react';
 import ImgRegularIcon from '../../assets/icons/kg-img-regular.svg?react';
 import ImgWideIcon from '../../assets/icons/kg-img-wide.svg?react';
 import React from 'react';
-import {ButtonGroupBeta, ButtonGroupIconButton} from './ButtonGroupBeta';
+import {ButtonGroup, IconButton} from './ButtonGroup';
 
 const story = {
-    title: 'Generic/Button group (beta)',
-    component: ButtonGroupBeta,
-    subcomponents: {ButtonGroupIconButton},
+    title: 'Generic/Button group',
+    component: ButtonGroup,
+    subcomponents: {IconButton},
     parameters: {
         status: {
             type: 'functional'
@@ -21,7 +21,7 @@ export default story;
 
 const Template = (args) => {
     return (
-        <ButtonGroupBeta {...args} />
+        <ButtonGroup {...args} />
     );
 };
