+++ packages/koenig-lexical/src/components/ui/MediaUploader_old.jsx
@@ -4,12 +4,12 @@ import PropTypes from 'prop-types';
 import WandIcon from '../../assets/icons/kg-wand.svg?react';
 import clsx from 'clsx';
 import {IconButton} from './IconButton';
-import {MediaPlaceholderBeta} from './MediaPlaceholderBeta';
+import {MediaPlaceholder} from './MediaPlaceholder';
 import {ProgressBar} from './ProgressBar';
 import {openFileSelection} from '../../utils/openFileSelection';
 import {useRef} from 'react';
 
-export function MediaUploaderBeta({
+export function MediaUploader({
     className,
     imgClassName,
     src,
@@ -17,7 +17,6 @@ export function MediaUploaderBeta({
     desc,
     icon,
     size,
-    type,
     borderStyle = 'squared',
     backgroundSize = 'cover',
     mimeTypes,
@@ -29,7 +28,7 @@ export function MediaUploaderBeta({
     openImageEditor,
     progress,
     errors,
-    onRemoveMedia = () => {},
+    onRemoveMedia,
     additionalActions,
     setFileInputRef
 }) {
@@ -54,7 +53,7 @@ export function MediaUploaderBeta({
     if (isEmpty) {
         return (
             <div className={className}>
-                <MediaPlaceholderBeta
+                <MediaPlaceholder
                     borderStyle={borderStyle}
                     dataTestId="media-upload-placeholder"
                     desc={isEditing ? desc : ''}
@@ -65,7 +64,6 @@ export function MediaUploaderBeta({
                     isDraggedOver={dragHandler?.isDraggedOver}
                     placeholderRef={dragHandler?.setRef}
                     size={size}
-                    type={type}
                 />
                 <ImageUploadForm
                     fileInputRef={onFileInputRef}
@@ -78,15 +76,10 @@ export function MediaUploaderBeta({
     }
 
     return (
-        <div className={clsx(
-            'group/image relative flex items-center justify-center', 
-            isLoading ? 'min-w-[6.8rem]' : 'min-w-[5.2rem]',
-            borderStyle === 'rounded' && 'rounded', 
-            className
-        )} data-testid="media-upload-filled">
+        <div className={clsx('group/image relative flex items-center justify-center', borderStyle === 'rounded' && 'rounded', className)} data-testid="media-upload-filled">
             {src && (
                 <>
-                    <img alt={alt} className={clsx('mx-auto h-full w-auto min-w-[5.2rem]', borderStyle === 'rounded' && 'rounded-lg', backgroundSize === 'cover' ? 'object-cover' : 'object-contain', imgClassName)} src={src} />
+                    <img alt={alt} className={clsx('mx-auto h-full w-auto', borderStyle === 'rounded' && 'rounded-lg', backgroundSize === 'cover' ? 'object-cover' : 'object-contain', imgClassName)} src={src} />
                     <div className={clsx('absolute inset-0 bg-gradient-to-t from-black/0 via-black/5 to-black/30 opacity-0 transition-all group-hover/image:opacity-100', borderStyle === 'rounded' && 'rounded-lg')}></div>
                 </>
             )}
@@ -110,7 +103,7 @@ export function MediaUploaderBeta({
 
             {isLoading && (
                 <div
-                    className={clsx('absolute inset-0 flex min-w-full items-center justify-center overflow-hidden bg-grey-100', borderStyle === 'rounded' && 'rounded-lg')}
+                    className={clsx('absolute inset-0 flex min-w-full items-center justify-center overflow-hidden border border-grey-100 bg-grey-75', borderStyle === 'rounded' && 'rounded-lg')}
                     data-testid="custom-thumbnail-progress"
                 >
                     <ProgressBar style={progressStyle} />
@@ -120,27 +113,24 @@ export function MediaUploaderBeta({
     );
 }
 
-MediaUploaderBeta.propTypes = {
-    additionalActions: PropTypes.node,
-    alt: PropTypes.string,
-    backgroundSize: PropTypes.oneOf(['cover', 'contain']),
-    borderStyle: PropTypes.oneOf(['squared', 'rounded']),
+MediaUploader.propTypes = {
     className: PropTypes.string,
+    src: PropTypes.string,
+    alt: PropTypes.string,
     desc: PropTypes.string,
-    dragHandler: PropTypes.shape({isDraggedOver: PropTypes.bool, setRef: PropTypes.func}),
-    errors: PropTypes.arrayOf(PropTypes.shape({message: PropTypes.string})),
     icon: PropTypes.string,
-    imgClassName: PropTypes.string,
-    isEditing: PropTypes.bool,
-    isLoading: PropTypes.bool,
-    isPinturaEnabled: PropTypes.bool,
+    size: PropTypes.string,
+    borderStyle: PropTypes.string,
     mimeTypes: PropTypes.arrayOf(PropTypes.string),
     onFileChange: PropTypes.func,
-    onRemoveMedia: PropTypes.func,
-    openImageEditor: PropTypes.func,
+    dragHandler: PropTypes.shape({
+        isDraggedOver: PropTypes.bool,
+        setRef: PropTypes.func
+    }),
+    isLoading: PropTypes.bool,
     progress: PropTypes.number,
-    setFileInputRef: PropTypes.func,
-    size: PropTypes.string,
-    src: PropTypes.string,
-    type: PropTypes.oneOf(['image', 'button'])
+    errors: PropTypes.arrayOf(PropTypes.shape({
+        message: PropTypes.string
+    })),
+    onRemoveMedia: PropTypes.func
 };
