+++ packages/koenig-lexical/src/components/ui/SettingsPanel.jsx
@@ -2,17 +2,17 @@ import KoenigComposerContext from '../../context/KoenigComposerContext.jsx';
 import React from 'react';
 import clsx from 'clsx';
 import useSettingsPanelReposition from '../../hooks/useSettingsPanelReposition';
-import {ButtonGroup} from './ButtonGroup';
-import {ButtonGroupBeta} from './ButtonGroupBeta';
-import {ColorIndicator, ColorPicker} from './ColorPicker';
-import {ColorIndicatorBeta} from './ColorPickerBeta';
-import {ColorOptionButtons} from './ColorOptionButtons';
-import {ColorOptionButtonsBeta} from './ColorOptionButtonsBeta.jsx';
+// import {ButtonGroup} from './ButtonGroup';
+import {ButtonGroup} from './ButtonGroup.jsx';
+// import {ColorIndicator, ColorPicker} from './ColorPicker';
+import {ColorIndicator} from './ColorPicker.jsx';
+// import {ColorOptionButtons} from './ColorOptionButtons';
+import {ColorOptionButtons} from './ColorOptionButtons.jsx';
 import {Dropdown} from './Dropdown';
 import {Input} from './Input';
 import {InputList, InputListItem} from './InputList.jsx';
-import {MediaUploader} from './MediaUploader';
-import {MediaUploaderBeta} from './MediaUploaderBeta';
+// import {MediaUploader} from './MediaUploader';
+import {MediaUploader} from './MediaUploader.jsx';
 import {MultiSelectDropdown} from './MultiSelectDropdown';
 import {Slider} from './Slider.jsx';
 import {TabView} from './TabView';
@@ -223,29 +223,41 @@ export function MultiSelectDropdownSetting({label, description, placeholder = ''
     );
 }
 
-export function ButtonGroupSetting({label, onClick, selectedName, buttons}) {
+// export function ButtonGroupSetting({label, onClick, selectedName, buttons}) {
+//     return (
+//         <div className="flex w-full items-center justify-between text-[1.3rem]">
+//             <div className="text-sm font-medium tracking-normal text-grey-900 dark:text-grey-300">{label}</div>
+
+//             <div className="shrink-0 pl-2">
+//                 <ButtonGroup buttons={buttons} selectedName={selectedName} onClick={onClick} />
+//             </div>
+//         </div>
+//     );
+// }
+
+export function ButtonGroupSetting({label, onClick, selectedName, buttons, hasTooltip}) {
     return (
         <div className="flex w-full items-center justify-between text-[1.3rem]">
             <div className="text-sm font-medium tracking-normal text-grey-900 dark:text-grey-300">{label}</div>
 
             <div className="shrink-0 pl-2">
-                <ButtonGroup buttons={buttons} selectedName={selectedName} onClick={onClick} />
+                <ButtonGroup buttons={buttons} hasTooltip={hasTooltip} selectedName={selectedName} onClick={onClick} />
             </div>
         </div>
     );
 }
 
-export function ButtonGroupSettingBeta({label, onClick, selectedName, buttons, hasTooltip}) {
-    return (
-        <div className="flex w-full items-center justify-between text-[1.3rem]">
-            <div className="text-sm font-medium tracking-normal text-grey-900 dark:text-grey-300">{label}</div>
+// export function ColorOptionSetting({label, onClick, selectedName, buttons, layout, dataTestId}) {
+//     return (
+//         <div className={`flex w-full text-[1.3rem] ${layout === 'stacked' ? 'flex-col' : 'items-center justify-between'}`} data-testid={dataTestId}>
+//             <div className="text-sm font-medium tracking-normal text-grey-900 dark:text-grey-300">{label}</div>
 
-            <div className="shrink-0 pl-2">
-                <ButtonGroupBeta buttons={buttons} hasTooltip={hasTooltip} selectedName={selectedName} onClick={onClick} />
-            </div>
-        </div>
-    );
-}
+//             <div className={`shrink-0 ${layout === 'stacked' ? '-mx-1 pt-[.6rem]' : 'pl-2'}`}>
+//                 <ColorOptionButtons buttons={buttons} selectedName={selectedName} onClick={onClick} />
+//             </div>
+//         </div>
+//     );
+// }
 
 export function ColorOptionSetting({label, onClick, selectedName, buttons, layout, dataTestId}) {
     return (
@@ -259,40 +271,53 @@ export function ColorOptionSetting({label, onClick, selectedName, buttons, layou
     );
 }
 
-export function ColorOptionSettingBeta({label, onClick, selectedName, buttons, layout, dataTestId}) {
-    return (
-        <div className={`flex w-full text-[1.3rem] ${layout === 'stacked' ? 'flex-col' : 'items-center justify-between'}`} data-testid={dataTestId}>
-            <div className="text-sm font-medium tracking-normal text-grey-900 dark:text-grey-300">{label}</div>
+// export function ColorPickerSetting({label, isExpanded, onSwatchChange, onPickerChange, onTogglePicker, value, swatches, eyedropper, hasTransparentOption, dataTestId}) {
+//     const mappedPicker = (event) => {
+//         onTogglePicker(true);
+//     };
 
-            <div className={`shrink-0 ${layout === 'stacked' ? '-mx-1 pt-[.6rem]' : 'pl-2'}`}>
-                <ColorOptionButtonsBeta buttons={buttons} selectedName={selectedName} onClick={onClick} />
-            </div>
-        </div>
-    );
-}
+//     const markClickedInside = (event) => {
+//         event.stopPropagation();
+//     };
 
-export function ColorPickerSetting({label, isExpanded, onSwatchChange, onPickerChange, onTogglePicker, value, swatches, eyedropper, hasTransparentOption, dataTestId}) {
-    const mappedPicker = (event) => {
-        onTogglePicker(true);
-    };
+//     // Close on click outside
+//     React.useEffect(() => {
+//         if (isExpanded) {
+//             const closePicker = (event) => {
+//                 onTogglePicker(false);
+//             };
+//             document.addEventListener('click', closePicker);
 
+//             return () => {
+//                 document.removeEventListener('click', closePicker);
+//             };
+//         }
+//     }, [isExpanded, onTogglePicker]);
+
+//     return (
+//         <div className="flex-col" data-testid={dataTestId} onClick={markClickedInside}>
+//             <div className="flex w-full items-center justify-between text-[1.3rem]">
+//                 <div className="text-sm font-medium tracking-normal text-grey-900 dark:text-grey-300">{label}</div>
+
+//                 <div className="shrink-0 pl-2">
+//                     <ColorIndicator
+//                         isExpanded={isExpanded}
+//                         swatches={swatches}
+//                         value={value}
+//                         onSwatchChange={onSwatchChange}
+//                         onTogglePicker={mappedPicker}
+//                     />
+//                 </div>
+//             </div>
+//             {isExpanded && <ColorPicker eyedropper={eyedropper} hasTransparentOption={hasTransparentOption} value={value} onChange={onPickerChange} />}
+//         </div>
+//     );
+// }
+
+export function ColorPickerSetting({label, isExpanded, onSwatchChange, onPickerChange, onTogglePicker, value, swatches, eyedropper, hasTransparentOption, dataTestId, customToolbarContent, children}) {
     const markClickedInside = (event) => {
         event.stopPropagation();
     };
-
-    // Close on click outside
-    React.useEffect(() => {
-        if (isExpanded) {
-            const closePicker = (event) => {
-                onTogglePicker(false);
-            };
-            document.addEventListener('click', closePicker);
-
-            return () => {
-                document.removeEventListener('click', closePicker);
-            };
-        }
-    }, [isExpanded, onTogglePicker]);
 
     return (
         <div className="flex-col" data-testid={dataTestId} onClick={markClickedInside}>
@@ -301,31 +326,6 @@ export function ColorPickerSetting({label, isExpanded, onSwatchChange, onPickerC
 
                 <div className="shrink-0 pl-2">
                     <ColorIndicator
-                        isExpanded={isExpanded}
-                        swatches={swatches}
-                        value={value}
-                        onSwatchChange={onSwatchChange}
-                        onTogglePicker={mappedPicker}
-                    />
-                </div>
-            </div>
-            {isExpanded && <ColorPicker eyedropper={eyedropper} hasTransparentOption={hasTransparentOption} value={value} onChange={onPickerChange} />}
-        </div>
-    );
-}
-
-export function ColorPickerSettingBeta({label, isExpanded, onSwatchChange, onPickerChange, onTogglePicker, value, swatches, eyedropper, hasTransparentOption, dataTestId, customToolbarContent, children}) {
-    const markClickedInside = (event) => {
-        event.stopPropagation();
-    };
-
-    return (
-        <div className="flex-col" data-testid={dataTestId} onClick={markClickedInside}>
-            <div className="flex w-full items-center justify-between text-[1.3rem]">
-                <div className="text-sm font-medium tracking-normal text-grey-900 dark:text-grey-300">{label}</div>
-
-                <div className="shrink-0 pl-2">
-                    <ColorIndicatorBeta
                         eyedropper={eyedropper}
                         hasTransparentOption={hasTransparentOption}
                         isExpanded={isExpanded}
@@ -336,12 +336,46 @@ export function ColorPickerSettingBeta({label, isExpanded, onSwatchChange, onPic
                         onTogglePicker={onTogglePicker}
                     >
                         {children}
-                    </ColorIndicatorBeta>
+                    </ColorIndicator>
                 </div>
             </div>
         </div>
     );
 }
+
+// export function MediaUploadSetting({className, label, hideLabel, onFileChange, isDraggedOver, placeholderRef, src, alt, isLoading, errors = [], progress, onRemoveMedia, icon, desc, size, type, stacked, borderStyle, mimeTypes, isPinturaEnabled, openImageEditor, setFileInputRef}) {
+//     return (
+//         <div className={clsx(className, !stacked && 'flex justify-between gap-3')} data-testid="media-upload-setting">
+//             <div className={hideLabel ? 'sr-only' : 'mb-2 shrink-0 text-sm font-medium tracking-normal text-grey-900 dark:text-grey-400'}>{label}</div>
+
+//             <MediaUploader
+//                 alt={alt}
+//                 borderStyle={borderStyle}
+//                 className={clsx(
+//                     stacked && 'h-32',
+//                     !stacked && src && 'h-[5.2rem]',
+//                     !stacked && type !== 'button' && !src && 'h-[5.2rem] w-[7.2rem]'
+//                 )}
+//                 desc={desc}
+//                 dragHandler={{isDraggedOver, setRef: placeholderRef}}
+//                 errors={errors}
+//                 icon={icon}
+//                 imgClassName={clsx(stacked && 'w-full')}
+//                 isLoading={isLoading}
+//                 isPinturaEnabled={isPinturaEnabled}
+//                 mimeTypes={mimeTypes}
+//                 openImageEditor={openImageEditor}
+//                 progress={progress}
+//                 setFileInputRef={setFileInputRef}
+//                 size={size}
+//                 src={src}
+//                 type={type}
+//                 onFileChange={onFileChange}
+//                 onRemoveMedia={onRemoveMedia}
+//             />
+//         </div>
+//     );
+// }
 
 export function MediaUploadSetting({className, label, hideLabel, onFileChange, isDraggedOver, placeholderRef, src, alt, isLoading, errors = [], progress, onRemoveMedia, icon, desc, size, type, stacked, borderStyle, mimeTypes, isPinturaEnabled, openImageEditor, setFileInputRef}) {
     return (
@@ -349,40 +383,6 @@ export function MediaUploadSetting({className, label, hideLabel, onFileChange, i
             <div className={hideLabel ? 'sr-only' : 'mb-2 shrink-0 text-sm font-medium tracking-normal text-grey-900 dark:text-grey-400'}>{label}</div>
 
             <MediaUploader
-                alt={alt}
-                borderStyle={borderStyle}
-                className={clsx(
-                    stacked && 'h-32',
-                    !stacked && src && 'h-[5.2rem]',
-                    !stacked && type !== 'button' && !src && 'h-[5.2rem] w-[7.2rem]'
-                )}
-                desc={desc}
-                dragHandler={{isDraggedOver, setRef: placeholderRef}}
-                errors={errors}
-                icon={icon}
-                imgClassName={clsx(stacked && 'w-full')}
-                isLoading={isLoading}
-                isPinturaEnabled={isPinturaEnabled}
-                mimeTypes={mimeTypes}
-                openImageEditor={openImageEditor}
-                progress={progress}
-                setFileInputRef={setFileInputRef}
-                size={size}
-                src={src}
-                type={type}
-                onFileChange={onFileChange}
-                onRemoveMedia={onRemoveMedia}
-            />
-        </div>
-    );
-}
-
-export function MediaUploadSettingBeta({className, label, hideLabel, onFileChange, isDraggedOver, placeholderRef, src, alt, isLoading, errors = [], progress, onRemoveMedia, icon, desc, size, type, stacked, borderStyle, mimeTypes, isPinturaEnabled, openImageEditor, setFileInputRef}) {
-    return (
-        <div className={clsx(className, !stacked && 'flex justify-between gap-3')} data-testid="media-upload-setting">
-            <div className={hideLabel ? 'sr-only' : 'mb-2 shrink-0 text-sm font-medium tracking-normal text-grey-900 dark:text-grey-400'}>{label}</div>
-
-            <MediaUploaderBeta
                 alt={alt}
                 borderStyle={borderStyle}
                 className={clsx(
