+++ packages/koenig-lexical/src/components/ui/cards/CallToActionCard.jsx
@@ -7,7 +7,7 @@ import ReplacementStringsPlugin from '../../../plugins/ReplacementStringsPlugin.
 import clsx from 'clsx';
 import defaultTheme from '../../../themes/default.js';
 import {Button} from '../Button.jsx';
-import {ButtonGroupSettingBeta, ColorOptionSettingBeta, ColorPickerSettingBeta, InputSetting, InputUrlSetting, MediaUploadSettingBeta, SettingsPanel, ToggleSetting} from '../SettingsPanel.jsx';
+import {ButtonGroupSetting, ColorOptionSetting, ColorPickerSetting, InputSetting, InputUrlSetting, MediaUploadSetting, SettingsPanel, ToggleSetting} from '../SettingsPanel.jsx';
 import {ReadOnlyOverlay} from '../ReadOnlyOverlay.jsx';
 import {RestrictContentPlugin} from '../../../index.js';
 import {VisibilitySettings} from '../VisibilitySettings.jsx';
@@ -167,7 +167,7 @@ export function CallToActionCard({
                 onChange={updateHasSponsorLabel}
             />
             {/* Image setting */}
-            <MediaUploadSettingBeta
+            <MediaUploadSetting
                 alt='Image'
                 borderStyle={'rounded'}
                 desc='Upload'
@@ -215,7 +215,7 @@ export function CallToActionCard({
     const designSettings = (
         <>
             {/* Layout settings */}
-            <ButtonGroupSettingBeta
+            <ButtonGroupSetting
                 buttons={layoutOptions}
                 hasTooltip={false}
                 label='Layout'
@@ -223,7 +223,7 @@ export function CallToActionCard({
                 onClick={updateLayout}
             />
             {/* Color picker */}
-            <ColorOptionSettingBeta
+            <ColorOptionSetting
                 buttons={callToActionColorPicker}
                 dataTestId='cta-background-color-picker'
                 label='Background'
@@ -231,7 +231,7 @@ export function CallToActionCard({
                 onClick={handleColorChange}
             />
             {/* Link color setting */}
-            <ColorOptionSettingBeta
+            <ColorOptionSetting
                 buttons={callToActionLinkColorPicker}
                 dataTestId='cta-link-color-picker'
                 label='Link color'
@@ -239,7 +239,7 @@ export function CallToActionCard({
                 onClick={handleLinkColorChange}
             />
             {showButton && (
-                <ColorPickerSettingBeta
+                <ColorPickerSetting
                     dataTestId='cta-button-color'
                     eyedropper={true}
                     isExpanded={buttonColorPickerExpanded}
